% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_doc.R
\docType{data}
\name{GE2014}
\alias{GE2014}
\title{New Zealand General Election Results 2014}
\format{A data frame with 136,195 rows and 7 variables.}
\source{
\link{http://www.electionresults.govt.nz/electionresults_2014/e9/html/statistics.html}
}
\usage{
GE2014
}
\description{
New Zealand 2014 General Election results by voting place
}
\details{
\itemize{
  \item ApproxLocation. Approximate location of voting place
  \item VotingPlace. Exact location of voting place
  \item Party. Party voted for (party vote) or party of candidate voted for
  (candidate vote)
  \item Votes. Number of votes
  \item Electorate. Electorate in which voters were enrolled.  Note that this 
  is not necessarily the physical location of the VotingPlace, so VotingPlace
  and Electorate have a many to many relationship
  \item VotingType. Party (proportional representation) or Candidate (first past 
  the vote).  In New Zealand each voter has to vote for both an individual
  candidate to represent their electorate, and a party vote for the overall makeup
  of Parliament.
  \item Candidate. If VotingType == "Candidate", the name of the candidate; 
  otherwise NA
}
}
\seealso{
\code{\link{XXXX}}
}
\keyword{datasets}

