% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_doc.R
\docType{data}
\name{Locations2014}
\alias{Locations2014}
\title{General Election Voting Places 2014}
\format{A data frame with 2,568 rows and 9 variables.}
\source{
\link{http://www.electionresults.govt.nz/electionresults_2014/e9/html/statistics.html}
}
\usage{
Locations2014
}
\description{
Voting places for the New Zealand 2014 general election
}
\details{
\itemize{
  \item ElectorateNumber.
  \item ElectorateName.  Name of electorate in which the voting place is 
  physically located.
  \item VotingPlaceID.
  \item NZTM2000Northing
  \item NZTM2000Easting
  \item WGS84Latitude
  \item WGS84Longitude
  \item VotingPlace.  Address of the VotingPlace.  Cross-references to
   GE2014$VotingPlace.
}
}
\examples{
# the following matches the results published at
# http://www.electionresults.govt.nz/electionresults_2014/e9/html/e9_part1.html
library(tidyr)
library(dplyr)
GE2014 \%>\%
    mutate(VotingType = paste0(VotingType, "Vote")) \%>\%
    group_by(Party, VotingType) \%>\%
    summarise(Votes = sum(Votes)) \%>\%
    spread(VotingType, Votes) \%>\%
    select(Party, PartyVote, CandidateVote) \%>\%
    ungroup() \%>\%
    arrange(desc(PartyVote))
}
\seealso{
\code{\link{GE2014}}
}
\keyword{datasets}

